pyproject.toml
[tool.poetry]
name = "python-bpmn-engine"
version = "0.1.0"
description = ""
authors = ["zyx <zyx1006@163.com>"]
readme = "README.md"
packages = [{include = "python_bpmn_engine"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
tortoise-orm = "^0.20.0"
requests = "^2.31.0"
uvicorn = "^0.29.0"
click = "^8.1.7"
h11 = "^0.14.0"
pony = "^0.7.17"
aiohttp = "^3.9.3"
aiohttp-cors = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
